Moved apollo client to base app dir
